Skip leave/join if unnecessary.
If the user already is a member of a group or is already removed from a
group, we should not be throwing an error, we should be skipping. This
will lookup a member in a group and not attempt to insert/remove that
member if they exist/don't respectively. The status of a skipped
insert/removal should be reflected in trogdir.